Decoupled command list from device
