+++ taichi/rhi/vulkan/CMakeLists.txt
@@ -10,6 +10,7 @@ target_sources(${VULKAN_RHI}
     vulkan_event.cpp
     vulkan_pipeline.cpp
     vulkan_resource_binder.cpp
+    vulkan_command_list.cpp
     vulkan_surface.cpp
     vulkan_stream.cpp
     vulkan_device_creator.cpp
